+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @callstack/brownfield-cli
 
+## 5.1.0
+
 ## 5.0.2
 
 ### Patch Changes